requirements.txt
# Core Dependencies
RPi.GPIO==0.7.1
pyusb==1.2.1
pyserial==3.5
spidev==3.5

# Web and API
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
websockets==11.0.3

# Data Processing
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1

# Hardware Interface
python-uinput==0.11.2

# Security
cryptography==40.0.2
paramiko==3.1.0

# Utilities
python-dotenv==1.0.0
rich==13.3.5
PyYAML==6.0

# Note: These versions are tested and confirmed working on Pi Zero W/2W
# Do not update versions without thorough testing
